+++ bdikit/models/contrastive_learning/cl_api.py
@@ -14,12 +14,13 @@ from sklearn.metrics.pairwise import cosine_similarity
 from tqdm.auto import tqdm
 from bdikit.download import get_cached_model_or_download
 from bdikit.utils import check_gdc_cache, write_embeddings_to_cache
+from bdikit.models import ColumnEmbedder
 
 
 DEFAULT_CL_MODEL = "bdi-cl-v0.2"
 
 
-class ContrastiveLearningAPI:
+class ContrastiveLearningAPI(ColumnEmbedder):
     def __init__(
         self,
         model_path: Optional[str] = None,